make: Add separate targets for building frontend, backend and tools
There was no way to build a frontend using `make` before, so adding
that now.

Also, rollerd is technically a backend, so now it can be built with
the `make backend`.

Other go binaries are basically tools, so let's use a separate target
for those.

Also update the Dockerfile accordingly.