rollerd: Drop useless update password functionality
With an introduction of the team-based authentication using github, a
notion of changing a password does not make any sense now - if you
want to change the password, do it on github.

This change is rather superficial as it only drops a body of a
function, so it becomes a no op. A proper change would involve
dropping the `/api/password` endpoint and changing the frontend so it
does not expose the functionality anymore.