rollerd: Use team-based oauth2 authentication using github
This drops the basic authentication used before. The outcome of this
is that rollerd requires passing client ID and client secret to it,
either through environment or by flags. Handling of the oauth
authentication required adding support for sessions and this adds an
optional flag to rollerd for passing a sessions secret.

For the authentication to work, coreroller teams should be named after
github organizations and teams. If the coreroller team is named just
like the github organization then any member of that organization will
be able to manage the applications of the equally named coreroller
team.

If the coreroller team is named after the github organization and team
(like "kinvolk/codemonkeys" where "kinvolk" is a name of github
organization and "codemonkeys" is a name of a team in that
organization) then only members of that specific github team will be
able to manage applications of the coreroller team.

Currently the code does not handle a scenario when user with a session
in rollerd is dropped from the github team. This is normally handled
through webhooks and will be added in later commits.

This also breaks getting metrics by prometheus, but will also be fixed
in a later commit.