rollerd: Add initial webhook functionality
This adds another piece of information that needs to be passed to
rollerd either through the environment or by a flag. Webhook secret is
used for computing the signature of the webhook message and comparing
it to the value of the X-Hub-Signature HTTP header. This is to make
sure that we handle only the legitimate messages sent by github.

This also adds tracking of sessions and teams, since webhook message
handles will essentially destroy sessions of the affected users or of
users in the affected teams. To be able to do it, we track all the
session IDs of a user and the team information that served as a base
of the successful authentication. We also track all the users using a
team. With that we can easily figure out all the sessions of some user
or some team. Currently we handle no messages and thus we don't use
any of the session and team tracking, but we will do it in later
commits.