rollerd: Handle "membership" event
Handling of that event is a bit similar to handling of the
"organization" event, but this time it is about teams.

We are only interested in the "removed" action in the "team" scope,
which means that some user was dropped from some team. When that
happens we try to find all the sessions of the user that were created
after sucessful authentication based on that team only and then
destroy them.