cmd/rollerd: set user access level to ro/rw
This commit drops the current mapping of github teams to nebraska teams
and rather assumes that there is only one nebraska team in the
database.

It uses the list of teams received from the -rw-teams and -ro-teams CLI
flags to set the access level of the user accordingly so that users in
read-only teams can only do GET and HEAD requests.

Signed-off-by: Ashijeet Acharya <ashijeet@kinvolk.io>