style: Replace Less CSS by SCSS
SCSS is the common way to preprocess CSS these days, and many js modules
used in Nebraska moved to it. So these changes translate the Less files
to SCSS.

It also loads the bootstrap and react-select stylesheets separately
from the project's one. This is because those projects still use Less
and this way we can move iteratively, without having to update the
stylesheets and the module in one go.