frontend: Update React to 16.8.6 and add compatibility changes
We want to update the frontend in order to use more recent/maintained
modules as the ones used previously are now quite old.

Since all the dependency modules used were quite old, dividing this
update in atomic commits proved to be very tricky. Hence, this is a big
patch as, in order to ensure a minimum compatibility and functionality
with the mentioned version of React, a number of modules also had to be
updated and the source code modified to address those changes.

Namely, the biggest changes are:
- react-router: it no longer accepts named routes or nested routes, so
  a lot has been modified to address this;
- react-spinners: this module replaces halogen very easily since it's
  API compatible and thus just requires an import replacement;
- legacy/react-bootstrap-validator: the original project hasn't been
  updated for 4 years and was not compatible with recent versions of
  react-bootstrap, so it's been removed from the dependencies and added
  as a legacy module after some changes;
- legacy/react-bootstrap: a lot of the forms both in Nebraska and also
  in react-bootstrap-validator use widgets from react-bootstrap that no
  longer exists; so this legacy module brings those back and changes
  them to ensure compatibility with newer modules too.

The reason for the changes above is to avoid introducing even more
changes in an already very big patch.

In terms of functionality, most of what was working before should be
working fine, but there's no assurance of visual/pixel compatibility
since that would require even more work and most of the bootstrap code
is to be replaced anyway.