frontend: Update use of React's PropTypes
The property types for React have been moved out of React itself and
into their own module (PropTypes). The way they should be assigned has
also changed (instead of a static member, they are now assigned outside
of the class definition).

This has not been a big problem, but when changing to a more restrictive
Babel, there are complaints about how the properties are built, so this
patch fixes this for every module that is using React's PropTypes.