frontend: Port/Refactor Applications edit/add forms
Instead of relying on Bootstrap + the legacy react-bootstrap-validation
modules, this patch changes how the edition and creation of applications
items is implemented by not only using Formik + Yup + Material UI but
also using only one form for both actions. This hence reduces the code
needed to maintain.

Since this is the first time porting a form to use Formik and Yup (for
the validation), this patch also introduces the mentioned dependencies
(as well as formik-material-ui).