frontend: Drop rc-switch dependency
This patch replaces the use of the rc-switch module and drops it from
the package.json.
For now, the switches are replaced simply by text as it was difficult
to understand whether the switcehes were enabled or not before anyway.

During the re-styling of the project, the replacement text may be
changed to a different component.