frontend: Make Channels/Item usable from different components
For consistency we should use the Channels/Item component not only in
the Channels section of the group page but also from the
Application/Group views.

This involves making the handleUpdateChannel property not manadatory as
well as allowing other properties to be passed to it (so it can be
e.g. displayed as something else than a "li" element).