frontend: Add a build tool to create iconify-compatible icons
Iconify offers a nice way to easily integrate icons from external
sources, thus, for consistency, we should make custom icons also
compatible with iconify.
In order to accomplish that, a json format file needs to be created for
the icon, so it is a good idea to have an automated process for this.

Thus, this patch adds a tool for creating those icons when building
with node (i.e. the icons json files will be generated when running
"make frontend").