frontend: Use the ChannelAvatar component for the color picker
The ColorPicker component used a colored button that did not resemble
the channel's avatr that it is pointing to. This patch changes that
and uses the ChannelAvatar component so it will have the same visuals
as the channel item displayed in lists.
It also uses an IconButton instead of a normal Button so it's rendered
as a circle instead of a square.