frontend: Allow to render different components as the ColorPicker
The ColorPicker uses the ChannelAvatar as the default renderer for the
color part of the button. While this is fine in many cases, there will
be others where we should display a different widget instead. So this
patch allows that by using whatever component is passed as the
ColorPicker's child, and by specifying a "componentColorProp" property
for the assigning the current color.

This functionality will be used in a later commit.