frontend: Add Update Progress section to the group's detailed view
It is useful to see the instances' status in terms of overall
percentage, so this patch adds a new sectin to the group's detailed view
with a circle progress chart for each interesting state.

This circle chart is implemented using the Victory library, so it's
added as a new dependency.