frontend: Fix first use of InstancesStore.getInstances
If the InstancesStore.getInstances method is used for the first time
and has a defined selectedInstance argument, then it will result in an
issue since it assumes that there is already an instances cache with
the selectedInstance and this is obviously not the case when calling it
for the first time.

With this patch, it checks if the selectedInstance actually exists
before attempting to use it, which solves the problem.