frontend: Add Instance details view
The current instances table view for the extended group view may be a
bit cumbersome to use, so this patch adds a new instance details view.
This new view offers a quick summary of an instance's details (ID, IP,
application, group, status, ...) and a "live" table with the latest
status' entries.