frontend: Refactor Instances' charts with makeStatusDefs
Since the Instance details view added a makeStatusDefs which holds the
information for the names, icons, and colors of the instances status,
then it is a good idea to use it in the Charts components too. This way,
we just have to touch one place if we want to change things.