pkg/api/db/migrations: add "default team ID"
The Nebraska database layout is based on Coreroller's and has an initial
design for "users" and "teams". But we don't use it in our current auth
implementation: instead, users are authenticated through GitHub OAuth
and all users are a member of the same "default team" with the same ID.
For different teams, different Nebraska instances should be used.

This team ID is expected to be available in the database during login.
If the ID cannot be read from the DB, it results in an internal server
error. Hence add the ID during the initial migration in order to avoid
errors during login with a freshly installed Nebraska instance (with no
data in the DB).