Add pkg/api/db/migrations/0006_initial_application.sql
This migration adds a "Flatcar Linux" application to Nebraska and
configures the default channels and groups. It also adds a package to
each channel to make sure a package is available from the start and
before the syncer has finished its first run (while the syncer is
started immediately as per next commit, it needs a few minutes to
actually update the packages).

Users who wish to not have a "Flatcar Linux" application but configure
their own can simply delete it from the UI.

Resolves #71