frontend: Extract the instances count label into its own module
For the upcoming instance list view we need to use an instance count
label that will be similar/equal to the one already used in the instance
status charts. Thus, this patch extracts that label into its own module,
and changes the name for the count text property in order to be more
accurate about its role.