frontend: Make instances selected item logic less interdependent
The instances list logic for the selected item was being treated in
different components for fetching the instance status history, i.e.
the item itself would call for the history to be loaded, but it was
the container two levels above it that would re-load the whole list
items in order for the instance status history to be effectively
shown.

With these changes, the item that gets selected and loads the instance
history is the one that gets the updated history and pass it down to
the status history view. This reduces the arquitecture complexity for
this component, and should also be faster.