frontend: Add new Instances page
The instance view is a new page for displaying the list of instances
that belong to a group.

The motivation for this page is that the group extended view (where the
instances table was displayed) was too crowded and unresponsive when
the instances were more than a couple of hundreds.

Thus, besides adding a completely new view for the instances, this patch
also removes the container component that was used before (in the group
extended view) and replaces it with a link to the new page.

In the new view, the table is paginated, which ensures that performance
is not an issue even when having thousands of instances.