frontend: Display several values in the Instances/Charts
The instances charts was displaying the data only related to 6 of the
instance status. This meant that the percentage displayed in those
charts was not summing up to 100, besides not informing the user of the
complete statuses of the instances.
A solution for this that consists of just adding more charts will not
work visually as it breaks the visual balance of having 6 charts, and
adds too much noise.

This patch fixes that issue by making the ProgressDoughnut chart accept
and display multiple values, and then passing more than one instance
status to some of the charts.
Namely, the "on-hold" and "undefined" statuses are now displayed
together (and the previously "On Hold" chart renamed as "Other"); the
"updated granted" and "downloading" status are also displayed together
the latter's chart (because "update granted" means that a download is
imminent).