frontend: Add a description to the Instances/Charts
Since now we are showing more than one instance status per
ProgressDoughnut chart, it is useful for the user to know exactly the
number of machine in each status.
To accomplish that, this patch adds a description to each status entry
in the charts, and displays that in a tooltip when the related chart's
are portion is hovered.

Similarly, when the "remaining" percentage or the chart's label is
hovered, the tooltip will show the description for all the values in it.