backend: Add API to fetch the version count timeline
We would like to display a chart that shows the timeline for the version
count, i.e. the number of instances per version on a timeline, as this
should give a good overview of the version update evolution during a
period.

This patch adds the API for that to the backend (only). The needed
frontend bits and the chart itself will be done in different commits.