frontend: Allow to set the format in makeLocaleTime
This function is supposed to help setting the right locale format to the
time, but some use-cases may need to specify what format the time should
have, e.g. only the day of the week, only the time, etc.

So this patch allows callers to specify the date and the format
independently.