frontend: Add the version count timeline to the group details
We want a way for the user to see the version evolution throughout a
period of time, so this patch adds a new interactive chart and table
showing the version count for the last 24 hours.

Initially the chart was implemented using Victory charts, but it was not
performing well enough, so a different charts library called Recharts
was used in the end for the implementation. Thus, this new dependency is
also added by the patch.