frontend: Remove the version progress bar from the group details
The version progress bar is not very useful in the group details page
since we now have the version count timeline. It also looks visually
very different, so this patch removes it from the page.

Since the section title was also contained in the progress bar's wrapper
component, this patch also adds a header to the section.