frontend: Separate and rename the InstanceChartSection
The InstanceChartSection was its own section (paper + header), but this
gives little flexibility if we want to extend it with elements already
in the page where the section was being displayed.

Thus, this patch separates the section part out of that component and
adds it instead to the ItemExtended page. Since the mentioned component
is no longer a section, this patch also renames it to
InstanceStatusArea.