frontend: Simplify the version breakdown progress bar
Each group item in the application view displays a progress bar with the
versions breakdown. However, this was implemented using the Nivo library
without much customization so it appeared visually very different from
the rest of the charts, as well as a bit too big and distracting.

This patch simplifies that by completely reimplementing the version
breakdown progress bar (using Recharts now), setting the colors
consistently with the ones used in the timeline chart, and moving it
under the "rollout policy" entry in that section.

Since the @nivo/bar package was added just for the previous progress
bar, this patch removes that dependency as well.