frontend: Add pagination to Packages/List
The Packages/List component was showing all the available packages for
an application. When the number of packages was high, this meant that
rendering the list took a while, and it didn't look visually good
either.

Thus, this patch limits the packages being displayed to 10 and adds a
pagination to the end of the packages section if the user needs to
display beyond the initial 10 packages.