backend: Fix recent data-addition migrations
The data addition migrations should have been done only if there was
no data already in the database. Otherwise if will violate the unique
key constraint and thus fail for existing deployments.

These changes make the mentioned data insertion dependent on whether
there is already a team and application, respective to what needs to
be inserted.