frontend: Add pagination to the activity list
The activity list can be very extense and thus increase the size of the
page beyond what's usable. To prevent that, these changes add a
pagination footer to the activity list.
This limitation on the activity list's size also improves the
performance of that page by not having to render a large number of list
items.