frontend: Fix z-index for the version progress bar
The version progress bar had the same z-index value as the tooltips
since it's tooltip should be displayed on top of the other elements (and
that was not the case with the default value).
However, this value was higher than e.g. the dialogs', which meant that
when editing a package, the progress bar would still be rendered above
the dialog.

This patch uses the "drawer" z-index value for the progress bar. This is
the same as the dialogs' and means that the progress bar is rendered on
top of the normal elements, but not on top of the dialogs.