backend: Add API to fetch the status count timeline
We would like to display a chart that shows the timeline for the version
count per status, i.e. the number of instances each version has for a
certain status and displayed on a timeline.
This should give a good overview of how many instances are reporting
status changes and what the overall status looks like over time.

This patch adds the API for that to the backend (only). The needed
frontend bits and the chart itself will be done in different commits.