frontend: Make the VersionStatsTable more generic
We will soon need to use a table just like the VersionStatsTable, so
instead of creating another table with hardcoded fields for a specific
dataset, this patch makes the VersionStatsTable more generic. It allows
for an empty message, the columns, and the value names to be defined.
This way we can reuse the code for this table.