frontend: Don't expect the channel to exist in the instances views
The Details and Tables views of the Instances module were expecting the
channel to always be valid when displaying its channel item. However, it
may be that the instance's group doesn't have a channel assigned to it,
in which case we should just show that information to the user (instead
of crashing and burning).