frontend: Ensure that percentage is a float in VersionBreakdownBar
There was an issue in the VersionBreakdownBar because the values of
certain percentages were being taken as string instead of a float (which
resulted in an error when attempting to fix its precision at a later
point).

This patch prevents that by ensuring that the percentage values are
always parsed as float before assigning/using them.