backend: Don't return invalid statuses/versions for timeline calls
The default values for the statuses/versions in timeline calls are
semantically invalid (e.g. 0 for a status, and "" for a version), so
this patch skips adding those.

The frontend was recently patched to tolerate those invalid values but
if we don't return them in the first place it also means avoiding extra
processing.