backend: Avoid panics when there are concurrent sessions
We are using sessions.MemoryStore as a way to store the session
data. MemoryStore is just an alias for a map type, so it seems to be
unsafe to use in case of concurrent sessions - we were getting panics
from runtime about concurrent map write. Arguably the library
providing the sessions middleware should be fixed by providing some
locking around session data operations, but we can work it around by
providing our own locking memory store.