frontend: Don't use a typographic wrapper in Packages/Item
By default, a ListItemtext wraps the primary and secondary properties
passed it to in a typography element. Since the component we are passing
as the secondary one was a Grid (defaults to "div"), which shouldn't be
added as a child of a "p" element (wrapper one), a warning about it was
being issued.
This patch adds the "disableTypography" property which stops the item
from using a typographic wrapper.