frontend: Make the Channels/Item render as a div
The Channels/Item uses a ListItem which renders to an "li" element
by default. This means that warnings are generated if it is wrapped
by an element that is not a list.

This patch fixes that by making the mentioned component render as a
"div", and removes the previous/wrong attempt of doing this by using
the ContainerComponent property of the ListItem.