build: Use go modules unconditionally
Even if the project is placed inside GOPATH, we still want to use
modules. This is not really necessary since go 1.13, but go 1.12 was
not using modules for projects inside GOPATH unless explicitly told to
do it with the GO111MODULES environment variable. We still use go 1.12
in Dockerfile.nebraska.