backend: Migrate to coreos/go-omaha
We migrate it, since the coreos version seems to be more up to date.

This was mostly a mechanical proces - some fields in structs have
different names (Id vs ID, or IsDelta vs IsDeltaPayload), some types
have changed (app status is a string-based enum, not a plain string
anymore), app response is now a different type from app request, or a
struct layout has changed a bit (first action in manifest as
manifest.Actions.Actions[0] vs manifest.Actions[0]).

What might be a kinda sorta of an incompatible change is that the
ChromeOSVersion field in the Action struct got renamed to
DisplayVersion.