backend: Pin the version of bindata
Different developers are going to have different versions of bindata
installed on their systems (either older, newer or coming from a
different repository - there are a lot of go-bindata forks out
there). This is only going to result in pull requests containing also
irrelevant changes in the regenerated files.

This adds github.com/kevinburke/go-bindata to go.mod at a pinned
version, so the regenerated file will be consistent. This version was
picked over the previously used twitter one, because the twitter
version seems to be unmaintained.

This commit still doesn't make pinned go-bindata to be actually used.