docker: Add dockerfile for postgres testing database
The created docker image will be used during CI tests. The entrypoint
script is a simplified version of a script for other postgres docker
image - we don't need to care about SSL setup so it's dropped.

The wait.sh script will be used by CI to wait with running the tests
until the database is ready.

The dockerfile is also a simplified version of the dockerfile for
other postgres image - we use su-exec (available as a package in
alpine) instead of gosu.