backend: Allow skipping tests completely
As a part of `make backend` we will want to run a linter too. The
linter will also check the sources of the test files. The problem is
that golangci-lint outputs unreadable messages in case of some syntax
errors. So we try building everything before linter is run, since the
compiler gives a much better output. Another problem is that there is
no way to just build the tests in all the packages. To work around
that, we specify the TestMain function in every package that contain
the tests, that returns early doing nothing if a special environment
variable is specified.