backend: Update golangci-lint rules
We add also goimports and golint checks to golangci-lint. This finds
some issues in the sources, so fix them immediately.

We will also stop running `gofmt -s` on the generated bindata file -
golangci-lint will do it for us. There seem to be some differences how
the formatting is done by golangci-lint, so apply those differences
too.