build: Add rules for CI
In CI, we want to perform the build of the project (which means
building, linting and checking go modules) and make sure that after
that, the git tree is still clean.

The CI will also run the tests using the testing postgres docker image
we added in an earlier commit.