backend: Conditionally repeat a grant of an update of an app
If we already granted an update of an app, but the client didn't act
on it yet (which means that it didn't send any event about starting
the downloading of the update binary, for example) and the client asks
for an update again then just resend the information about the granted
update.

One test needed an update for this case.