frontend: Add support for architecture in Packages/EditDialog
When creating a new package manually, it is required to choose an
architecture for it. So this patch adds an architecture field to the
mentioned form, which is used not only to choose the architecture
when creating the package, but to display which one the package has
when editing it (it's not allowed to modify the architecture of an
existing package).