frontend: Filter the packages shown in the Channels/EditDialog
We should only show packages with the same architecture as the channel.
So this patch filters the packages it exposes to the user in that form's
list. When creating a new channel, the filtering is dynamic, i.e. it
changes according to what the user has selected as the channel's
architecture in that moment.