frontend: Prevent showing a user icon in the channel's avatar
When no text is assigned to the channel's avatar, it gets assigned a
user icon. This wasn't noticeable before for some reason, but now it's
clear in the package items (where the channel's avatar is tiny and has
no initial assigned to it).

This patch makes the channel avatar use an empty string by default when
no contents are given to it.