backend: Refactor router setup
Some of the routing setup will be done by the future authenticator
type and we want to force the authenticator to set up the routing in
the same way as nebraska does for all its routes (which currently is
about assigning the names to routers, so we can log used routers in a
request).

Authenticator will live in a separate package, which is why we created
the small ginhelpers package: to avoid cyclic imports. The main
package imports the ginhelpers package to implement the
ginhelpers.Router interface and to set up routes, and the
authenticator package will import the ginhelpers package to set up its
additional routes. Without the ginhelpers package we would end up with
the authenticator package importing main package to use the Router
interface, and the main package importing the authenticator package to
use the authenticators.