backend: Add a "main" function that returns an error
To return from the main function with some non-zero exit status, one
must use `os.Exit(…)`. This is getting repetitive if we also want to
log the error that caused us to exit prematurely.

Just add another kinda sorta main function that returns an error, so
we can log it and exit with a status indicating failure in a single
place.