frontend: Prevent multiple calls to getInstances when listing them
In the instances list view, due to how the side-effect is being used,
getting the list of instances was happening multiple times and that is
unnecessary.

So this patch avoids that by only getting the instances if the last
check happened more than a minute ago.

This is a quick fix for the problem mentioned above, since a more
complete and elegant solution would be to implement some time+changes
based verification in the instances store, but that requires more time
and we want to avoid this issue now.