frontend: Load all instances for the group being viewed
The instances list view was displaying only 500 instances because the
backend paginates the results using that number by default.

Making good use of the pagination itself is not possible for our use
case (where we need to display a total count of the filtered results)
because there is no current way of getting the total count by itself,
and the pagination is happening before the filtering (version/status)
is applied. See https://github.com/kinvolk/nebraska/issues/122 .