frontend: Don't show the loading state forever in instances list view
The list view was showing the loading state until at least 1 instance
was fetched. However, if the group has no instances, then it should
just say so.
This patch accomplishes that by defining the instances list state as
null until we signalled from the instance store, at which point the
state is an empty list, thus displaying the desired empty message.

Notice that the UI shouldn't allow to access the instances list view
for groups that have no instances, but the user can of course access
them directly, or have it opened and refresh it into such state.