backend: Add new client config REST endpoint
This endpoint can be used to tell clients about important configuration
such as the access management endpoint (which depends on the
authentication backend and configuration).

For now, only the access management endpoint is set up by the backend.
If Github is used as the auth backend, it sets the public or the
enterprise settings URL depending on which version of Github is being
used.