frontend: Use the access amangement URL given by the backend
The access management URL (accessible from the user menu of the header)
was hardcoded from when Github was the only authentication backend that
Nebraska supported.

Currently, Nebraska supports a "noop" authentication backend for
testing, as well as the Github and Github Enterprise ones, and all of
them have different access management URLs.
Thus, in order to reflect the right authentication as mentioned, this
patch gets the access management URL from the backend as part of its
new "client configuration", and uses it in the respective link (or
hides the menu altogether if the URL is empty, as with the "noop"
backend).

These changes also make the access management menu future-proof (when other backend options are added).