frontend: Make showing the arch optional in Channels/Item
There are several improvements to be done in the way that the channels
and packages are showing with different architectures. One of the plans
is to group the channel items per architecture when listing them, and in
that case they shouldn't display their architecture as part of the
secondary text (to avoid redundancy/clutter).

Hence, this patch adds an option to the Channels/Item component that
makes it show/hide their architecture as needed.