docs: Update/fix the recommended way of running the database
The documentation about running a database for development was still
using the container built by CoreRoller (and with the name updated to
Nebraska by mistake).

Instead of relying on a dedicated container for the database, we should
just provide the instructions to quickly set it up for Nebraska. So
these changes replace the mentioned legacy instructions this way.