frontend: Fix instance history view when filter status changed
When showing an instance's history in the instances list, if the list
was filtered, then a re-render would happen and the instance for which
the history was being shown could end up in a different page.

This happened because the filteredInstances was being used inside
useEffect but it was not mentioned in its change variables array. This
patch adds it to that list and hence solves the problem.