frontend: Fix Dev Environment
As Earlier we were simply building the frontend and serving it from
backend and because of which we were not able to use features like hot
reloading or having a seperate server to host dev for frontend.

This Patch adds many cool features like hot module replacement,live
reloading also the directory structure is made very much similar to that
of create-react-app.