frontend: Replace moment.js by the standard Date API
We were using moment for date and time handling but the problem with
using moment is that it is a very large package and thus was increasing
our javascript bundle size. So this patch removes moment.js and uses
the javascript Date API this way our bundle size will drastically
decrease which is a good thing as now our javascript will load faster.