backend: Correctly generate package URL for download
The update package URL was being generated by simply concatenating the
base URL with the package's name, and this could create an invalid URL
if the base URL didn't have an ending slash.

Instead of relying on the existence of that ending slash, this patch
fixes the issue mentioned above by joining the URL with the url and
path modules, which ensure that a slash is added when needed.