frontend: Add eslint to project
We were not having a linter to our project till date and this was
leading to lot of different code patterns(like somebody prefering let
over const) from different contributors with eslint we will have a
global config which brings everyone on the same page.
This patch adds eslint to the CRA bootstrapped project also we need to
eject the create-react-app in order to override the lint configurations.