frontend: Fix eslint setup
We were earlier using react-app-rewired which was using the .eslintrc
for the linter, but this resulted in build errors and we want to avoid
that.Hence, this patch removes the react-app-rewired dependency, as well
as the use of the default create-react-app lint rules.
The idea is that we just use the .eslintrc file provided directly
(i.e. outside of the build process).