frontend: Fix timelineInterval constant to pass query without spaces
Earlier we were passing query params with spaces which led to encoding
of spaces in url. This patch fixes timeInterval constant to now have
queryValues without spaces and also adds another field called
displayValue which will be used to display link names to user. This
patch fixes the constant as well as updates all the places where the
constant was used earlier.