frontend: Fix race condition for api calls on duration change
So the bug is if you select one duration than goes an api call and till
the api response is fetched if you move to another duration tab than
goes api call for that tab but the thing is render component for each
duration tab is the same so there is a race condition of api calls and
the api which gets prior to other renders its data this creates a lot of
inconsistency. This patch fixes the race condition.