backend: Allow getApp to accept duration query param.
In order to fetch instance_stats duration wise we need to change the
getApp api to accept duration as query param and filter instance_stats
based on duration passed. If no duration is passed we fallback to
default duration i.e., 1 days. This patch adds the logic to accept
duration param and act accordingly.