frontend: Fix fetch call to handle backend response statuses
Our API request to the backend returns a status if something goes wrong,
e.g. when the server is down or 404. These status needs to be handled on
the frontend because some of the API calls like getInstanceStatusHistory
uses the status 404 to set specific data to some property. This patch
updates the fetch calls to handle statuses before making a call to
response.json() so that if the request is not OK, it just throws that
error.