frontend: Paginate the instances when fetching them
As we now have pagination working in the backend when fetching
instances, we can use this feature for making the listing of instances
less heavy, i.e we can fetch the instances already featured and
paginated, instead of getting them all and paginating them in the
frontend.

So this patch adds the mentioned changes, which makes client more
performant, and simplify the whole logic involved.