backend: Fix fetching the instances with pagination
Fetching the instances with pagination and filters was broken since it
was paginating before applying the filters. The pagination was also made
more difficult for the client because the backend call it was not
returning the total number of instances (filtered), so the client
couldn't know whether it was in the last page or not.

These changes fix both issues mentioned above, i.e. they apply the
filters before the pagination, and return the total number of instances
as well.