build: Rename Dockerfile.nebraska to Dockerfile
We used to have Dockerfile.nebraska and Dockerfile.postgres, for
distinguishing between the two major images of the project. However, we
have now removed the Postgres one as its not being used, so we should
also rename the Nebraska one as simply Dockerfile to emphasize it is
the main image (we still have Dockerfile.postgres-test, but that's
clearly a helper image).