frontend: Add proxy for /config on dev server
As our frontend and backend run on two different server in dev mode and
we proxy our /api to backend server there was one more thing that needs
to be proxied to backend and that is /config (this call provides the
client configs). Thus this patch adds /config to the proxy.