frontend: Add filter and pagination to url
When we apply filters and change pagination there was no way we can
remember what filters were selected when someones visits to instance
extended view from instance list view this lead to loosing filters
and pagination when somebody visits back to the instance list view.
Thus this patch adds filters and pagination info to url by pushing them
to router history this way we dont loose filters when we come back to
the instance list view.